feat: update how create page modal is rendered
