+++ packages/hash/frontend/src/components/layout/MainContentWrapper.tsx
@@ -1,13 +1,8 @@
-import {
-  faChevronLeft,
-  faChevronRight,
-} from "@fortawesome/free-solid-svg-icons";
 import { IconButton, Fade, Box, Tooltip, styled } from "@mui/material";
 import { FunctionComponent } from "react";
 import { SIDEBAR_WIDTH } from "../../theme/components/navigation/MuiDrawerThemeOptions";
-import { Button } from "../Button";
 
-import { FontAwesomeSvgIcon, SidebarToggleIcon } from "../icons";
+import { SidebarToggleIcon } from "../icons";
 import { HEADER_HEIGHT } from "./PageHeader/PageHeader";
 import { PageSidebar } from "./PageSidebar/PageSidebar";
 import { useSidebarContext } from "./SidebarContext";
@@ -74,169 +69,7 @@ export const MainContentWrapper: FunctionComponent = ({ children }) => {
           </IconButton>
         </Tooltip>
       </Fade>
-      <Main sidebarOpen={sidebarOpen}>
-        {children}
-        <Box mb={6} />
-        <Box sx={{ display: "flex", alignItems: "flex-start" }}>
-          <Button variant="primary" size="large">
-            Primary Large
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button size="medium">Primary Medium Button</Button>
-          <Box sx={{ mr: 2 }} />
-          <Button size="small">Primary Small Button</Button>
-          <Box sx={{ mr: 2 }} />
-          <Button size="xs">Primary XS Button</Button>
-          <Box sx={{ mr: 2 }} />
-        </Box>
-        <br />
-        <Box sx={{ display: "flex", alignItems: "flex-start" }}>
-          <Button variant="secondary" size="large">
-            Secondary Large
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="secondary" size="medium">
-            Secondary Medium Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="secondary" size="small">
-            Secondary Small Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="secondary" size="xs">
-            Secondary XS Button
-          </Button>
-        </Box>
-        <br />
-        <Box sx={{ display: "flex", alignItems: "flex-start" }}>
-          <Button variant="tertiary" size="large">
-            Tertiary Large
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="tertiary" size="medium">
-            Tertiary Medium Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="tertiary" size="small">
-            Tertiary Small Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="tertiary" size="xs">
-            Tertiary XS Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-        </Box>
-        <br />
-        <Box sx={{ display: "flex", alignItems: "flex-start" }}>
-          <Button variant="tertiary_quiet" size="large">
-            Tertiary Quiet Large
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="tertiary_quiet" size="medium">
-            Tertiary Quiet Medium Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="tertiary_quiet" size="small">
-            Tertiary Quiet Small Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="tertiary_quiet" size="xs">
-            Tertiary Quiet XS Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-        </Box>
-        <br />
-        <Box sx={{ display: "flex", alignItems: "flex-start" }}>
-          <Button variant="warning" size="large">
-            Warning Large
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="warning" size="medium">
-            Warning
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="warning" size="small">
-            Warning Small Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="warning" size="xs">
-            Warning XS
-          </Button>
-          <Box sx={{ mr: 2 }} />
-        </Box>
-        <br />
-        <Box sx={{ display: "flex", alignItems: "flex-start" }}>
-          <Button variant="danger" size="large">
-            Delete Large
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="danger" size="medium">
-            Delete
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="danger" size="small">
-            Delete Small Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="danger" size="xs">
-            Delete XS
-          </Button>
-          <Box sx={{ mr: 2 }} />
-        </Box>
-        <br />
-        <Box sx={{ display: "flex", alignItems: "flex-start" }}>
-          <Button disabled>Disabled Button</Button>
-          <Box sx={{ mr: 2 }} />
-          <Button disabled size="small">
-            {" "}
-            Disabled button{" "}
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button
-            disabled
-            disabledTooltipText="Please enter a title or type to continue"
-          >
-            A disabled button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-        </Box>
-        <br />
-        <Box sx={{ display: "flex", alignItems: "flex-start" }}>
-          <Button loading>Primary</Button>
-          <Box sx={{ mr: 2 }} />
-          <Button variant="secondary" size="medium" loadingWithoutText>
-            Secondary
-          </Button>
-          <Box sx={{ mr: 2 }} />
-        </Box>
-        <br />
-        <Box sx={{ display: "flex", alignItems: "flex-start" }}>
-          <Button
-            size="small"
-            href="/"
-            startIcon={<FontAwesomeSvgIcon icon={faChevronLeft} />}
-          >
-            Button
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button
-            variant="secondary"
-            size="small"
-            endIcon={<FontAwesomeSvgIcon icon={faChevronRight} />}
-          >
-            Secondary
-          </Button>
-          <Box sx={{ mr: 2 }} />
-          <Button
-            variant="secondary"
-            size="small"
-            href="/"
-            startIcon={<FontAwesomeSvgIcon icon={faChevronLeft} />}
-          >
-            Button
-          </Button>
-        </Box>
-      </Main>
+      <Main sidebarOpen={sidebarOpen}>{children}</Main>
     </Box>
   );
 };
