+++ packages/hash/frontend/src/components/layout/PageSidebar/AccountEntityTypeList/AccountEntityTypeList.tsx
@@ -1,14 +1,16 @@
 import { useMemo, VFC } from "react";
 
-import { Typography, Box } from "@mui/material";
+import { Typography, Box, Tooltip, IconButton } from "@mui/material";
 // import { faSearch, faArrowUpAZ } from "@fortawesome/free-solid-svg-icons";
 // import { orderBy } from "lodash";
+import { useRouter } from "next/router";
+import { faAdd } from "@fortawesome/free-solid-svg-icons";
 import { useAccountEntityTypes } from "../../../hooks/useAccountEntityTypes";
 // import { FontAwesomeSvgIcon } from "../../../icons";
 import { NavLink } from "../NavLink";
 import { Link } from "../../../Link";
 import { EntityTypeMenu } from "./EntityTypeMenu";
-import { CreateEntityTypeButton } from "./CreateEntityTypeButton";
+import { FontAwesomeSvgIcon } from "../../../icons";
 
 type AccountEntityTypeListProps = {
   accountId: string;
@@ -18,6 +20,7 @@ export const AccountEntityTypeList: VFC<AccountEntityTypeListProps> = ({
   accountId,
 }) => {
   const { data } = useAccountEntityTypes(accountId);
+  const router = useRouter();
   //   const [order, setOrder] = useState<"asc" | "desc" | undefined>();
 
   const sortedData = useMemo(() => {
@@ -54,7 +57,19 @@ export const AccountEntityTypeList: VFC<AccountEntityTypeListProps> = ({
     <Box>
       <NavLink
         title="Types"
-        endAdornment={<CreateEntityTypeButton accountId={accountId} />}
+        endAdornment={
+          <Tooltip title="Create new type">
+            {/* @todo-mui use a LinkButton here once it has been implemented */}
+            <IconButton
+              data-testid="create-entity-btn"
+              onClick={() => {
+                void router.push(`/${accountId}/types/new`);
+              }}
+            >
+              <FontAwesomeSvgIcon icon={faAdd} />
+            </IconButton>
+          </Tooltip>
+        }
       >
         <Box component="ul">
           {/* Can be uncommented once we have a page that displays all entity types */}
